requirements.txt
aeronet==0.0.18
affine==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==0.4.6
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cssutils==2.6.0
cycler==0.11.0
dataframe-image==0.1.11
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
h5py==3.8.0
html2image==2.0.3
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.5.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.2.1
multidict==6.0.4
munch==2.5.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.2.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
rasterio==1.3.6
requests==2.28.1
Rtree==1.0.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
segmentation-models-pytorch==0.3.2
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tifffile==2023.4.12
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
